Updates to dshop backend

+++ dapps/shop/backend/routes/auth.js
@@ -1,44 +1,50 @@
 const omit = require('lodash/omit')
 const { Sellers, Shops } = require('../data/db')
-const {
-  createSalt,
-  hashPassword,
-  AuthSeller,
-  authenticatedAsSeller
-} = require('./_combinedAuth')
+const { checkPassword } = require('./_auth')
+const { createSeller } = require('../utils/sellers')
 const { IS_PROD } = require('../utils/const')
 const encConf = require('../utils/encryptedConfig')
 const { validateConfig, validateShop } = require('../utils/validators')
 
 module.exports = function(app) {
   app.get('/auth', (req, res) => {
-    res.json({ success: req.isAuthenticated() })
+    if (!req.session.sellerId) {
+      return res.json({ success: false })
+    }
+    Sellers.findOne({ where: { id: req.session.sellerId } }).then(seller => {
+      res.json({ success: true, email: seller.email })
+    })
   })
 
   app.get('/auth/:email', async (req, res) => {
     // TODO: Add some rate limiting here
-    const seller = await Sellers.findOne({
-      where: {
-        email: req.params.email
-      }
-    })
+    const { email } = req.params
+    const seller = await Sellers.findOne({ where: { email } })
     return res.sendStatus(seller === null ? 404 : 204)
   })
 
-  app.post(
-    '/auth/login',
-    (req, res, next) => {
-      return next()
-    },
-    AuthSeller,
-    (req, res) => {
-      res.json({ success: req.isAuthenticated() })
-    }
-  )
+  app.post('/auth/login', async (req, res) => {
+    Sellers.findOne({ where: { email: req.body.email } }).then(seller => {
+      if (!seller) {
+        res.status(404).send({ success: false })
+        return
+      }
+      if (checkPassword(seller.password, req.body.password)) {
+        req.session.sellerId = seller.id
+        res.json({ success: true })
+      } else {
+        res.json({ success: false })
+      }
+    })
+  })
 
   const logoutHandler = (req, res) => {
-    if (req.isAuthenticated()) req.logout()
-    res.json({ success: !req.isAuthenticated() })
+    if (req.session.sellerId) {
+      req.logout()
+      res.json({ success: true })
+    } else {
+      res.json({ success: false })
+    }
   }
 
   app.get('/auth/logout', logoutHandler)
@@ -46,71 +52,45 @@ module.exports = function(app) {
 
   // TODO: Should this at least use API key auth?
   app.post('/auth/registration', async (req, res) => {
-    const { name, email, password } = req.body
-    if (!name || !email || !password) {
-      return res.status(400).json({
-        success: false,
-        message: 'Invalid registration'
-      })
+    const { seller, status, error } = createSeller(req.body)
+
+    if (error) {
+      return res.status(status).json({ success: false, message: error })
     }
-
-    const sellerCheck = await Sellers.findOne({
-      where: {
-        email
-      }
-    })
-
-    if (sellerCheck) {
-      return res.status(409).json({
-        success: false,
-        message: 'Registration exists'
-      })
-    }
-
-    const salt = await createSalt()
-    const passwordHash = await hashPassword(salt, password)
-
-    const seller = await Sellers.create({
-      name,
-      email,
-      password: passwordHash
-    })
 
     if (!seller) {
       return res.json({ success: false })
     }
 
-    return res.json({ success: true })
+    req.session.sellerId = seller.id
+    res.json({ success: true })
   })
 
-  app.delete('/auth/registration', authenticatedAsSeller, async (req, res) => {
-    const { id } = req.user
+  app.delete('/auth/registration', async (req, res) => {
+    const { sellerId } = req.session
 
-    if (!id) {
+    if (!sellerId) {
       return res.status(400).json({ success: false })
     }
 
-    const destroy = await Sellers.destroy({
-      where: {
-        id
-      }
-    })
+    if (!sellerId) {
+      return res.status(400).json({ success: false })
+    }
 
+    const destroy = await Sellers.destroy({ where: { id: sellerId } })
+
+    req.logout()
     res.json({ success: false, destroy })
   })
 
-  app.get('/shop', authenticatedAsSeller, async (req, res) => {
-    const { id } = req.user
+  app.get('/shop', async (req, res) => {
+    const { sellerId } = req.session
 
-    if (!id) {
+    if (!sellerId) {
       return res.status(400).json({ success: false })
     }
 
-    const rows = await Shops.findAll({
-      where: {
-        sellerId: id
-      }
-    })
+    const rows = await Shops.findAll({ where: { sellerId } })
 
     const shops = []
     for (const row of rows) {
@@ -125,7 +105,11 @@ module.exports = function(app) {
     })
   })
 
-  app.post('/shop', authenticatedAsSeller, async (req, res) => {
+  app.post('/shop', async (req, res) => {
+    if (!req.session.sellerId) {
+      return res.json({ success: false })
+    }
+
     const shopObj = req.body
 
     if (!validateShop(shopObj)) {
@@ -136,30 +120,32 @@ module.exports = function(app) {
 
     const shop = await Shops.create({
       ...shopObj,
-      sellerId: req.user.id
+      sellerId: req.session.sellerId
     })
 
     res.json({ success: true, shop })
   })
 
-  app.delete('/shop', authenticatedAsSeller, async (req, res) => {
+  app.delete('/shop', async (req, res) => {
+    if (!req.session.sellerId) {
+      return res.json({ success: false })
+    }
+
     await Shops.destroy({
       where: {
         id: req.body.id,
-        sellerId: req.user.id
+        sellerId: req.session.sellerId
       }
     })
     res.json({ success: true })
   })
 
-  app.post('/config', authenticatedAsSeller, async (req, res) => {
+  app.post('/config', async (req, res) => {
+    const { sellerId } = req.session
     const { shopId, config } = req.body
 
     const shop = await Shops.findOne({
-      where: {
-        id: shopId,
-        sellerId: req.user.id
-      }
+      where: { id: shopId, sellerId }
     })
 
     if (!shop) {
@@ -178,24 +164,21 @@ module.exports = function(app) {
     return res.json({ success: true })
   })
 
-  app.get('/config/dump/:id', authenticatedAsSeller, async (req, res) => {
-    const { id: shopId } = req.params
+  app.get('/config/dump/:id', async (req, res) => {
+    const { id } = req.params
 
     // Testing only
     if (IS_PROD) return res.sendStatus(404)
 
     const shop = await Shops.findOne({
-      where: {
-        id: shopId,
-        sellerId: req.user.id
-      }
+      where: { id, sellerId: req.session.sellerId }
     })
 
     if (!shop) {
       return res.status(404)
     }
 
-    const config = await encConf.dump(shopId)
+    const config = await encConf.dump(id)
     return res.json({ success: true, config })
   })
 }
